fix: TypeError: Cannot read property 'errorCount' of undefined
Without this patch, there are situations when the .solhintignore file
causes all inputs files being ignored. This is a problem because the
'reports' variable ends up being an empty array, and therefore
reports[0] is undefined.

This patch solves the problem by checking that reports array exists and
has a length greater than 0.

Thanks to Adam Spiers (@aspiers) who originally discovered a bug.

Issue-URL: https://bytemeta.vip/repo/protofire/solhint/issues/239