fix: bug directories with .sol in the name treated as files
Without this patch, the path, for example,
artifacts/@openzeppelin/contracts/token/ERC20/ERC20.sol/ERC20.dbg.js
will be treated as a file
artifacts/@openzeppelin/contracts/token/ERC20/ERC20.sol
on MacOS system. As a result, npm module fs throws the error:
'Error: EISDIR: illegal operation on a directory'

This patch solves the problem by checking if the path is a file and not
a directory.