no-global-import rule
disallow imports that potentially pollute the namespace by importing
every defined name of the imported file.

~when registering an ImportDirective method on the NoGlobalImportsChecker
class, it's only called by imports that explicitly define a new
Identifier, the only way I found to find global imports is to search for
them in the SourceUnit nodes~

^ was recommended against in review and when trying to replicate it to
show the issue to fvictorio, I wasn't able to. So the cleaner way, using
the ImportDirective visitor, will remain.